+++ executables/game/sources/game/gameActions/moveCurrentPieceToMatrix.cpp
@@ -1,24 +1,23 @@
-#include <array>
-#include "game/consts/constexpr.h"
-#include "matrices/matrix2D.h"
+#include "game/gameActions/moveCurrentPieceToMatrix.h"
 #include "game/structs/scanLinesInfos.h"
 #include "game/objects/randomSingleton.h"
 #include "game/objects/loadedPieces.h"
 #include "game/objects/activePieceThings.h"
-#include "game/gameActions/moveCurrentPieceToMatrix.h"
+#include "matrices/matrix2D.h"
+#include "game/consts/constexpr.h"
 
 void moveCurrentPieceToMatrix(const ActivePiece& pieceData, const Piece& piece, MatrixTemp2D<SquareData>& matrix, ScanLinesInfos& scanLinesInf)
 {
-	for( int posx{0} ; posx < piece.width(pieceData.rotateIndex) ; ++posx )
+	for( size_t posx{0} ; posx < piece.width(pieceData.rotateIndex) ; ++posx )
 	{
-		for( int posy{0} ; posy < piece.height(pieceData.rotateIndex) ; ++posy )
+		for( size_t posy{0} ; posy < piece.height(pieceData.rotateIndex) ; ++posy )
 		{
 			if( piece.getPresenceSquare(pieceData.rotateIndex, posx, posy) )
 			{
-				if( matrix.isSignedInsideBoundaries(pieceData.posx + posx, pieceData.posy + posy) )
+				if( matrix.isSignedInsideBoundaries(pieceData.posx + static_cast<int>(posx), pieceData.posy + static_cast<int>(posy) ) )
 				{
-					matrix(pieceData.posx + posx, pieceData.posy + posy).color = static_cast<unsigned>( pieceData.colorIndex );
-					matrix(pieceData.posx + posx, pieceData.posy + posy).filled = true;
+					matrix(pieceData.posx + static_cast<int>(posx), pieceData.posy + static_cast<int>(posy) ).color = static_cast<unsigned>( pieceData.colorIndex );
+					matrix(pieceData.posx + static_cast<int>(posx), pieceData.posy + static_cast<int>(posy) ).filled = true;
 				}
 			}
 		}
