+++ executables/game/sources/game/objects/activePieceThings.h
@@ -1,7 +1,7 @@
 #ifndef ACTIVE_PIECE_THINGS_H
 #define ACTIVE_PIECE_THINGS_H
 
-#include <cstddef>
+#include <cstdlib>
 
 class RandomSingleton;
 class LoadedPieces;
@@ -12,7 +12,7 @@ struct ActivePiece
 	size_t rotateIndex;
 	int posx;
 	int	posy;
-	size_t colorIndex;
+	unsigned colorIndex;
 	
 	ActivePiece(RandomSingleton& rand, const LoadedPieces& allPieces);
 	void resetPieceValues(RandomSingleton& rand, const LoadedPieces& pieces);
